Separate classes out into multiple files
