+++ .gitignore
@@ -100,3 +100,6 @@ venv.bak/
 
 # mypy
 .mypy_cache/
+
+.idea
+.vscode
