+++ uqcsbot/__main__.py
@@ -1,1 +1,2 @@
-from . import main; main()
+from . import main
+main()
